pyproject.toml
[tool.poetry]
name = "tempo-tracker"
version = "0.1.0"
description = ""
authors = ["Joe Wesch <joe.wesch@networktocode.com>"]

[tool.poetry.dependencies]
python = "^3.9"
requests = "^2.27.1"
Flask = "^2.1.2"
python-dateutil = "^2.8.2"
numpy = "^1.22.3"

[tool.poetry.dev-dependencies]
black = "^22.3.0"
prettytable = "^3.3.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
target-version = ['py37']
include = '\.pyi?$'
exclude = '''
(
  /(
      \.eggs         # exclude a few common directories in the
    | \.git          # root of the project
    | \.hg
    | \.mypy_cache
    | \.tox
    | \.venv
    | _build
    | buck-out
    | build
    | dist
  )/
  | settings.py     # This is where you define files that should not be stylized by black
                     # the root of the project
)
'''
